requirements.txt
fastapi
uvicorn

requests 
black
ipdb

# OTel
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests

# Sentry
-e ../sentry-python/